requirements.txt
asgiref==3.5.0
black==22.3.0
boto3==1.21.42
botocore==1.24.42
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Django==4.0.4
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==3.3
jmespath==1.0.0
mypy-extensions==0.4.3
numpy==1.22.3
openai==0.18.1
openpyxl==3.0.9
pandas==1.4.2
pandas-stubs==1.2.0.58
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tqdm==4.64.0
urllib3==1.26.9
